Move EmptyAddress and EmptyAdditionalInformation into their own namespace
